refactor : fileReader 삭제, AttendancesFile가 파일 읽어오도록 변경
