--- src/main/java/domain/FileReader.java
@@ -1,22 +1,0 @@
-package domain;
-
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Stream;
-
-public class FileReader {
-    public List<String> readLines(Path path) {
-        try (Stream<String> lines = Files.lines(path, StandardCharsets.UTF_8)) {
-            return lines.skip(1)
-                    .map(String::trim)
-                    .filter(line -> !line.isEmpty())
-                    .toList();
-        } catch (IOException e) {
-            return new ArrayList<>();
-        }
-    }
-}
